Single tool for MCP generation
All multiple steps of data processing are now governed from one place data-transformer.

The json-emitter was moved from the vim_build project. Single data model is shared in generation and mcp server runtime.

A fix was made to the mcp server path resolution to have lookup of top level properties in a type.

All data processing logging now uses the tracing crate

`cuda` feature is enabled in the consolidated data processing pipeline

All tools in the processing pipeline are modified to allow use as libraries.
