+++ mcp/.gitignore
@@ -1,0 +1,1 @@
+model_cache
