+++ mcp/data_processing/build_guides/Cargo.toml
@@ -8,3 +8,5 @@ anyhow = "1.0"
 regex = "1.10"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
+tracing = "0.1"
+tracing-subscriber = "0.3"
