+++ mcp/data_processing/pdf_parser/Cargo.toml
@@ -9,3 +9,5 @@ path = "src/main.rs"
 
 [dependencies]
 extractous = "0.3"
+tracing = "0.1"
+tracing-subscriber = "0.3"
