+++ mcp/data_processing/text_processor/Cargo.toml
@@ -11,3 +11,5 @@ path = "src/main.rs"
 anyhow = "1.0"
 regex = "1.11"
 lazy_static = "1.4"
+tracing = "0.1"
+tracing-subscriber = "0.3"
