+++ mcp/server/Cargo.toml
@@ -12,6 +12,7 @@ tracing = "0.1"
 tracing-subscriber = "0.3"
 phf = "0.11"
 thiserror = "1.0"
+chrono = { version = "0.4", features = ["serde"] }
 
 # MCP protocol implementation
 rmcp = { version = "0.8.5", features = ["server", "macros"] }
