+++ vim_build/src/lib.rs
@@ -1,0 +1,10 @@
+pub mod vim_model;
+pub mod generator;
+pub mod rs_emitter;
+pub mod printer;
+
+// Re-export commonly used items
+pub use generator::load_openapi;
+pub use vim_model::load_vim_model;
+pub use rs_emitter::names;
+
