+++ vim_macros/src/field_data.rs
@@ -13388,3 +13388,9 @@ pub(crate) fn lookup_field_data(class: &str, field: &str) -> Result<&'static Nod
             }
         })
 }
+#[allow(dead_code)]
+pub(crate) fn get_type_fields(class: &str) -> Result<&'static phf::Map<&'static str, NodeData>> {
+    CLASS_FIELDS
+        .get(class)
+        .ok_or_else(|| HierarchyError::UnsupportedObjectType(class.to_string()))
+}
